Move api key input to new settings page
This will make it easier to add more settings, and to navigate to the
settings screen programmatically.
