+++ src/router.ts
@@ -1,7 +1,8 @@
 import Vue from 'vue';
 import Router from 'vue-router';
-import Home from './views/Home.vue';
+import Home from './views/Home/Home.vue';
 import Locations from './views/Locations.vue';
+import Settings from './views/Settings/Settings.vue';
 
 Vue.use(Router);
 
@@ -16,6 +17,11 @@ export default new Router({
       path: '/locations',
       name: 'locations',
       component: Locations,
+    },
+    {
+      path: '/settings',
+      name: 'settings',
+      component: Settings,
     },
   ],
 });
